var: added new var TIMEOUT

+++ userbot/modules/help.py
@@ -5,7 +5,7 @@
 #
 """ Userbot help command """
 
-from userbot import CMD_HELP, HELP_TIMEOUT
+from userbot import CMD_HELP, TIMEOUT
 from userbot.events import register
 from asyncio import sleep
 @register(outgoing=True, pattern=r"^\.help(?: |$)(.*)")
@@ -16,29 +16,49 @@ async def help(event):
     if event.is_channel and not event.is_group:
         await event.edit("`Help command isn't permitted on channels`")
         return
-    if args:
-        if args in CMD_HELP:
-            msg=await event.edit(str(CMD_HELP[args]))
-            if HELP_TIMEOUT:
-                await sleep(45)
+    
+    if TIMEOUT:
+        if args:
+            if args in CMD_HELP:
+                msg=await event.edit(str(CMD_HELP[args]))
+                await sleep(10)
                 await msg.delete()
+                
             else:
-                return
+                msg=await event.edit("Please specify a valid module name.")
+                await sleep(10)
+                await msg.delete()
+                           
         else:
-            msg=await event.edit("Please specify a valid module name.")
-            await sleep(10)
-            await msg.delete()
-    else:
-        final = "**List of all loaded module(s)**\n\
+            final = "**List of all loaded module(s)**\n\
                  \nSpecify which module do you want help for! \
                  \n**Usage:** `.help` <module name>\n\n"
-
-        temp = "".join(str(i) + " " for i in CMD_HELP)
-        temp = sorted(temp.split())
-        for i in temp:
-            final += "`" + str(i)
-            final += "`\t\t\t•\t\t\t"
-        msg=await event.edit(f"{final[:-5]}")
-        if HELP_TIMEOUT:
-            await sleep(45)
-            await msg.delete()
+            temp = "".join(str(i) + " " for i in CMD_HELP)
+            temp = sorted(temp.split())
+            for i in temp:
+                final += "`" + str(i)
+                final += "`\t\t\t•\t\t\t"
+            msg=await event.edit(f"{final[:-5]}")
+            await sleep(10)
+            await msg.delete()   
+    
+    
+    if not TIMEOUT:
+        if args:
+            if args in CMD_HELP:
+                msg=await event.edit(str(CMD_HELP[args]))
+                
+            else:
+                msg=await event.edit("Please specify a valid module name.")
+                           
+        else:
+            final = "**List of all loaded module(s)**\n\
+                 \nSpecify which module do you want help for! \
+                 \n**Usage:** `.help` <module name>\n\n"
+            temp = "".join(str(i) + " " for i in CMD_HELP)
+            temp = sorted(temp.split())
+            for i in temp:
+                final += "`" + str(i)
+                final += "`\t\t\t•\t\t\t"
+            msg=await event.edit(f"{final[:-5]}")
+
